#noid #done #comment feature/back/page[이미지|Id타입|total]

See merge request s08-webmobile1-sub2/S08P12E104!162

+++ backend/src/main/java/com/ssafy/patpat/board/entity/Comment.java
@@ -19,11 +19,11 @@ import java.time.LocalDateTime;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int commentId;
+    private Long commentId;
     private String content;
     private LocalDateTime regTime;
-    private int boardId;
-    private int userId;
+    private Long boardId;
+    private Long userId;
     private String nickName;
 
     public void updateComment(String content){
